+++ outbound/index.js
@@ -241,7 +241,7 @@ exports.send_trans_email = function (transaction, next) {
         transaction.results = new ResultStore(connection)
     }
 
-    connection.pre_send_trans_email_respond = async (retval) => {
+    connection.pre_send_trans_email_respond = async () => {
         const deliveries = get_deliveries(transaction)
         const hmails = []
         const ok_paths = []